% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignTools-package.R
\docType{package}
\name{alignTools-package}
\alias{alignTools-package}
\alias{alignTools}
\title{alignTools}
\description{
easy download and manage GenBank data, and alignments for phylogenetics
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado12@gmail.com}{leon.alvarado12@gmail.com}
}
\keyword{packagelibrary}
