% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeGenes.R
\name{completeGenes}
\alias{completeGenes}
\title{Complete Alignments}
\usage{
completeGenes(
  genes,
  missing = c("?", "-"),
  write.dna = c(TRUE, FALSE),
  write.format = c("nexus", "phylip", "fasta", "tnt")
)
}
\arguments{
\item{genes}{A list object with multiple alignments (DNAbin class)}

\item{missing}{When a species or individual have no information for a gene, the absent gene should be filled with "?" or "-". by default "?".}

\item{write.dna}{Boolean, write on separate files each gene with the complete list of species or individuals. By default TRUE.}

\item{write.format}{The format to write each alignment (only if write.dna==TRUE). Four options: 1. nexus (default), 2. phylip, 3. fasta or 4. tnt.}
}
\value{
A list object with multiple alignments (DNAbin class). All with the same number of species or individuals.
}
\description{
Some softwares (e.g. RevBayes) for partitioned analysis required each gene in a separate file, but all files must have the same number of species. This function achive that task.
}
\examples{

library(alignTools)

data(Procyonidae)
head(Procyonidae, 5L)

genes <- multiGenBank(Procyonidae,TRUE,FALSE)

alignment <- multiMuscle(genes, FALSE)

complete <- completeGenes(genes=alignment, missing="?", write.dna=TRUE, write.format="fasta")

str(complete)

}
\seealso{
\code{\link{concatGenes}} \code{\link{read.multiGenes}}
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado12@gmail.com}{leon.alvarado12@gmail.com}
}
\keyword{alignments}
\keyword{complete}
\keyword{genes}
