% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatGenes.R
\name{concatGenes}
\alias{concatGenes}
\title{Concatenate Multiple Alignments}
\usage{
concatGenes(
  Genes,
  missing = c("?", "-"),
  write.dna = c(TRUE, FALSE),
  write.format = c("nexus", "phylip", "fasta", "tnt"),
  filename = NULL
)
}
\arguments{
\item{missing}{When a species or individual have no information for one or more genes, the absent gene(s) should be filled with "?" or "-". By default "?".}

\item{write.dna}{Boolean, write the concatenate alignment in a file with predeterminate format. By deafault TRUE.}

\item{write.format}{The format to write the concatenate alignment (only if write.dna == TRUE). Three options: 1. nexus (default), 2. phylip, 3. fasta or 4. tnt.}

\item{filename}{The name of the output file (only if write.dna==TRUE).}

\item{genes}{A list object with multiple alignments (DNAbin class)}
}
\value{
A concatenate alignment of DNAbin class
}
\description{
This function concatenate multiple alignments
}
\details{
\strong{Author:} Omar Daniel Leon-Alvarado, <leon.alvarado12@@gmail.com> \if{html}{\out{<br>}}
\strong{Repository:} \url{https://github.com/oleon12/alignTools}
\subsection{\strong{Install}}{\if{html}{\out{<div class="r">}}\preformatted{library(devtools)
install_github("oleon12/alignTools")
}\if{html}{\out{</div>}}
}

\subsection{\strong{Data}}{
\subsection{To download multiple genes from GenBank, the input file we use is a matrix. This matrix should have in the first column the species names or identifiers, and the others columns will be the genes and their accession numbers. So, this is the “hard” step, you need to find each accession number. When a species have not information for a gene, you fill it as NA. See the Procyonidae data.}{\if{html}{\out{<div class="r">}}\preformatted{#Procyonidae data
head(Procyonidae, 5L)
}\if{html}{\out{</div>}}\preformatted{##               Species     cytb      coi     rag2
## 1         Nasua_nasua KU253485 KF771221 GU931274
## 2        Nasua_narica DQ533940 DQ533934 DQ660275
## 3  Bassaricyon_alleni DQ660299     <NA> GU931277
## 4  Bassaricyon_gabbii JX948744 DQ533937 DQ660273
## 5 Bassaricyon_neblina EF107709     <NA>     <NA>
}
}

}

\subsection{\strong{multiGenBank( )}}{
\subsection{With this function, we will download the three genes for all species in Procyonidae. This function have two important parameters: 1. with.na: When you have NA in your data, you set it TRUE; 2. write.dna: If you want to save the files for each gene, you set it TRUE. The file will be save in “fasta” format and the name will be the same as in the columns. The result is a list object with 3 DNAbin objects. See that each DNAbin object have a name, this name is the same as the column name.}{\if{html}{\out{<div class="r">}}\preformatted{genes <- multiGenBank(x = Procyonidae, with.na = TRUE, write.dna = FALSE)
#Check genes, a list object
str(genes)
genes$cytb
}\if{html}{\out{</div>}}
}

}

\subsection{\strong{multiMuscle( )}}{
\subsection{Now the next step is to align our sequences. For this we will use multiMuscle(), this function will align the sequences of each gene. The input file must be a list object. Now, only the MUSCLE algorithm is available, in future I will update for others like Clustal or tcoffe. The result is practically the same list object, but each gene aligned. Like multiGenBank(), this function can save each alignment in separate files, to do, just set write.dna=TRUE and select one of the avaiable formats: “nexus” (default), “phylip”, “fasta” or “tnt”.}{\if{html}{\out{<div class="r">}}\preformatted{alignment <- multiMuscle(x = genes, write.dna = FALSE)
#check alignment, a list object
str(alignment)
alignment$coi
}\if{html}{\out{</div>}}
}

}

\subsection{\strong{concatGenes( )}}{
\subsection{Most softwares (e.g. MrBayes, RaxML, PhyML) require a single file of concatenate genes. With concatGenes() we will achieve this task. It is common that one o more species have not information for one o more genes, so, when concatenate, we need to fill this absence. With concantGenes() we have two options: “?”, or “-”, by default “?”. Also, we can save the concatenated file, just set write.dna=TRUE, select one format: “nexus” (default), “phylip”, “fasta” or “tnt”, and a filename. Very important, we can only concatenate aligned sequences.}{\if{html}{\out{<div class="r">}}\preformatted{concat <- concatGenes(Genes = alignment, missing = "?", write.dna = FALSE)
#Check concat, a DNAbin object
concat
}\if{html}{\out{</div>}}
}

}

\subsection{\strong{completeGenes( )}}{
\subsection{Same parameters as concatGenes(). Some softwares (e.g. RevBayes or PhyML) for partitioned analysis, require each gene in separate files, but all genes must have the same number of species. With completeGenes() we can create those files with the same number of species. For species without information in a specific gene, the function will use “?” (default) or “-”. The result is a list object with DNAbin objects, all of them with the same number of species. Also, we can save the concatenated file, just set write.dna=TRUE, and select one format: “nexus” (default), “phylip”, “fasta” or “tnt”. Very important, we can only use aligned sequences.}{\if{html}{\out{<div class="r">}}\preformatted{complete <- completeGenes(genes = alignment, missing = "?", write.dna = FALSE)
#Check complete, a list object
str(complete)
complete$rag2
}\if{html}{\out{</div>}}
}

}

\subsection{\strong{multiModelTest( )}}{
\subsection{This function is an extension of the modelTest() function from phangorn. So, detailed information can be found with \emph{help(modelTest)}. This function only works with aligned sequences. The results is a list object of length three. 1. summary.results: is a summary table with the best models for each gene with each estimator (AIC, AICc, AICw, AICcw and BIC), 2. results: The complete results of ModelTest(), and 3. warnings: sometimes, a gene could have two models, when that happend, a "*" symbol will be in the summary table, and the details will be saved in warnings.}{\if{html}{\out{<div class="r">}}\preformatted{models <- multiModelTest(genes = alignment, G = TRUE, I = TRUE)
#Check models, a list object
models$warnings
models$results$cytb
}\if{html}{\out{</div>}}
}

}

\subsection{\strong{read.multiGenes( )}}{
\subsection{If you already have your own data, and want to use some functions from this package (e.g. concatGenes() or multiMuscle()), you can read your files using read.multiGenes(), this function will create the appropiate list object with DNAbin objects. Remember that the files must be the in the same working directory (setwd()). If you want to set specific names to each file, use the “names” parameter, but be careful, the names order had to be the same files order. The function can read three formats: “nexus”, “phylip” or “fasta”.}{\if{html}{\out{<div class="r">}}\preformatted{files <- c("data1.nex","data2.nex","data3.nex","data4.nex")
#No names
data <- read.multiGenes(files = files, format = "nexus")
str(data)
data$data4.nex
#With names
data <- read.multiGenes(files = files, format = "nexus", names = c("gene1","gene2","gene3","gene4"))
str(data)
data$gene3
#Other way to set names
data <- read.multiGenes(files = files, format = "nexus")
names(data) <- c("my_name1","my_name2","my_name3","my_name4")
}\if{html}{\out{</div>}}
}

}

\subsection{\strong{partitionScheme( )}}{
\subsection{When we concatenate some genes, a partition scheme is created. This shows the genes order, and the position of each gene within the alignment. If we forgot this information, using partitionScheme() and the alignment, we can recover it. This could help you to set the partitions in the software (e.g. MrBayes, RaxML).}{\if{html}{\out{<div class="r">}}\preformatted{concat <- concatGenes(Genes = alignment, missing = "?", write.dna = FALSE)
#Check partition scheme
partitionScheme(alignment)
}\if{html}{\out{</div>}}
}

}

\subsection{\strong{Pipelines} \if{html}{\out{<br>}}}{
\subsection{Pipeline 1: Common concatenate file. (works for MrBayes, PhyML, RAxML and many others)}{\if{html}{\out{<div class="r">}}\preformatted{#Nexus
multiGenBank(Procyonidae,TRUE,FALSE)\%>\%
multiMuscle(write.dna = FALSE)\%>\%
concatGenes(missing = "?", write.dna = TRUE, write.format = "nexus", filename = "concat")
#Phylip
multiGenBank(Procyonidae,TRUE,FALSE)\%>\%
multiMuscle(write.dna = FALSE)\%>\%
concatGenes(missing = "?", write.dna = TRUE, write.format = "phylip", filename = "concat")
}\if{html}{\out{</div>}}
}

\subsection{Pipeline 2: Partitioned analysis with separate files (works for RevBayes and PhyML)}{\if{html}{\out{<div class="r">}}\preformatted{#Nexus
multiGenBank(Procyonidae, TRUE, FALSE)\%>\%
multiMuscle(write.dna = FALSE)\%>\%
completeGenes(missing = "?", write.dna = TRUE, write.format = "nexus")
#Phylip
multiGenBank(Procyonidae, TRUE, FALSE)\%>\%
multiMuscle(write.dna = FALSE)\%>\%
completeGenes(missing = "?", write.dna = TRUE, write.format = "phylip")
}\if{html}{\out{</div>}}
}

}
}
\examples{

library(alignTools)

data(Procyonidae)
head(Procyonidae, 5L)

genes <- multiGenBank(Procyonidae,TRUE,FALSE)

alignment <- multiMuscle(x=genes, FALSE)

concat <- concatGenes(genes=alignment, missing="?", write.dna=TRUE, write.format="phylip", filename="Concat_eg_output")

str(concat)

}
\seealso{
\code{\link{multiMuscle}} \code{\link{completeGenes}}
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado12@gmail.com}{leon.alvarado12@gmail.com}
}
\keyword{alignments}
\keyword{concatenate}
\keyword{genes}
