% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGenBank.R
\name{multiGenBank}
\alias{multiGenBank}
\title{Get Multiple Sequences from GenBank}
\usage{
multiGenBank(x, with.na = c("TRUE", "FALSE"), write.dna = c("TRUE", "FALSE"))
}
\arguments{
\item{x}{is the matrix or data frame with the access numbers of each gene for each species or individual. This matrix must have the species or indivual names in the first column. See example.}

\item{with.na}{Boolean, if there is any NA in the matrix, must be set as TRUE (default). When you have no data or access number in a gene for one or more species, you must codify this as NA, otherwise, the function will not work.}

\item{write.dna}{Boolean, write each gene as separate files in fasta format. The file names will be the same as the column names. By deafault TRUE.}
}
\value{
The function returns a list object with the sequences for each gene (DNAbin class). Even setting write.dna=TRUE, the function will returns the list object.
}
\description{
The function download the sequences for multiple genes. The sequences will be stored in a list object and/or will be saved in different files in fasta format.
}
\examples{

library(alignTools)

data(Procyonidae)
head(Procyonidae, 5L)

genes <- multiGenBank(x=Procyonidae, with.na=TRUE, write.dna=FALSE)
str(genes)
genes$cytb

}
\seealso{
\code{\link{multiMuscle}} \code{\link{read.multiGenes}}
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado12@gmail.com}{leon.alvarado12@gmail.com}
}
\keyword{GenBank}
\keyword{download}
\keyword{sequences}
