% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModelTest.R
\name{multiModelTest}
\alias{multiModelTest}
\title{Nucleotide Substitution Models for Multiple Alignments}
\usage{
multiModelTest(
  genes,
  tree = NULL,
  model = c("JC", "F81", "HKY", "K80", "SYM", "GTR"),
  G = TRUE,
  I = TRUE,
  FREQ = FALSE,
  k = 4,
  control = pml.control(epsilon = 1e-08, maxit = 10, trace = 1),
  multicore = FALSE,
  mc.cores = NULL
)
}
\arguments{
\item{genes}{A list objecti with multiple alignments (DNAbin class)}

\item{model}{The models to evaluate. By default JC, F81, HKY, K80, SYM and GTR. Or use "all".}

\item{G}{Boolean, Test Gamma. By default TRUE}

\item{I}{Boolean, Test invariant sites. By default TRUE}
}
\value{
A list object of length of three. 1. summary.results: A data frame with the selected models for each alignment and each metric (AIC, AICw, AICc, AICcw and BIC). 2. results: Detailed information for each alignment (list object), the normal output of modelTest. 3. warnings: When a alignment have two o more models selected, the warning shows the models selected for each alignment and each metric.
}
\description{
The function is an extension for multiple alignments of the modelTest() function from phangorn. For more information help("modelTest")
}
\examples{

library(alignTools)

data(Procyonidae)
head(PRocyonidae, 5L)

genes <- multiGenBank(Procyonidae, TRUE, FALSE)

alignment <- multiMuscle(genes, FALSE)

models <- multiModelTest(genes=alignment)

models$results.summary
str(models$results)
models$results$cytb
models$warnigs

}
\seealso{
\code{\link{multiMuscle}}
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado@gmail.com}{leon.alvarado@gmail.com}
}
\keyword{ModelTest}
\keyword{nucleotide}
\keyword{substitution}
