% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMuscle.R
\name{multiMuscle}
\alias{multiMuscle}
\title{Multiple Genes Alignment with MUSCLE}
\usage{
multiMuscle(
  x = NULL,
  y,
  guide.tree,
  quiet = TRUE,
  original.ordering = TRUE,
  MoreArgs = "",
  write.dna = c("TRUE", "FALSE"),
  write.format = c("nexus", "phylip", "fasta", "tnt")
)
}
\arguments{
\item{x}{A list object with multiple DNAbin objects}

\item{write.dna}{Boolean, write each alignment as separate files in a predeterminate format. The file names will be the same as the x object names. By deafault TRUE.}

\item{write.format}{The format to write each alignment (only if write.dna == TRUE). Four options: 1. nexus (Default), 2. phylip, 3. fasta or 4. tnt.}
}
\value{
A list object with multiple alignments (DNAbin class).
}
\description{
This function is an extesion the muscle() function from ape. The x parameter is the only changed here, so, for more information about the other parameters, please see the documentation from ape.
}
\examples{

library(alignTools)

data(Procyonidae)
head(Procyonidae, 5L)

genes <- multiGenBank(Procyonidae, TRUE, FALSE)

alignment <- multiMuscle(x=genes, write.dna=TRUE, write.format="phylip")

str(alignment) 
alignment$rag2

}
\seealso{
\code{\link{multiGenBank}}
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado12@gmail.com}{leon.alvarado12@gmail.com}
}
\keyword{alignment}
\keyword{genes}
\keyword{muscle}
