% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionScheme.R
\name{partitionScheme}
\alias{partitionScheme}
\title{Partition Scheme for a Concatenate Alignment}
\usage{
partitionScheme(genes)
}
\arguments{
\item{genes}{A list objecto with multiple alginments (DNAbin class)}
}
\description{
Shows a partition scheme for a concatenate alignment of N genes
}
\examples{

library(alignTools)

data(Procyonidae)
head(Procyonidae, 5L)

genes <- multiGenBank(Procyonidae,TRUE,FALSE)
alignment <- multiMuscle(genes, FALSE)

concat <- concatGenes(alignment, "?", FALSE)

partitionScheme(alignment)

}
\seealso{
\code{\link{concatGenes}} \code{\link{completeGenes}}
}
\author{
Omar Daniel Leon-Alvarado
}
\keyword{alignments}
\keyword{genes}
\keyword{partition}
\keyword{scheme}
