% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.multiGenes.R
\name{read.multiGenes}
\alias{read.multiGenes}
\title{Read Multiple Genes/Alignments files}
\usage{
read.multiGenes(files, format = c("nexus", "phylip", "fasta"), names = NULL)
}
\arguments{
\item{files}{Vecotr with the name of the files to read. The files must be in the same working directory.}

\item{format}{The format of the files. Three options: 1. nexus, 2. phylip or 3. fasta}

\item{names}{(optional), a vector with same lenght as files. This represent an specific name for each file. If missin, the function will use the files vector as names.}
}
\description{
The function reads two o more dna files (alignments)
}
\examples{

library(alignTools)

input <- c("files1.nex","file2.nex","file3.nex")

# In this example, names will be the same as input
data <- read.multiGenes(files=input, format="nexus")

# In this exmaple, names will be "id1","id2","id3".
data <- read.multiGenes(files=input, format="nexus",names=c("id1","id2","id3"))

str(data)

}
\seealso{
\code{\link{multiGenBank}} \code{\link{concatGenes}}
}
\author{
Omar Daniel Leon-Alvarado
}
\keyword{alignments}
\keyword{genes}
\keyword{multiple}
\keyword{read}
