% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt.dna.R
\name{tnt.dna}
\alias{tnt.dna}
\title{TNT format syntax (DNA)}
\usage{
tnt.dna(init, filename)
}
\arguments{
\item{init}{DNAbin object}

\item{filename}{output file name}
}
\description{
TNT format syntax (DNA)
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado12@gmail.com}{leon.alvarado12@gmail.com}
}
\keyword{DNA}
\keyword{TNT}
\keyword{write}
