% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatGenes.R
\name{concatGenes}
\alias{concatGenes}
\title{Concatenate Multiple Alignments}
\usage{
concatGenes(
  Genes,
  missing = c("?", "-"),
  write.dna = c(TRUE, FALSE),
  write.format = c("nexus", "phylip", "fasta", "tnt"),
  filename = NULL
)
}
\arguments{
\item{missing}{When a species or individual have no information for one or more genes, the absent gene(s) should be filled with "?" or "-". By default "?".}

\item{write.dna}{Boolean, write the concatenate alignment in a file with predeterminate format. By deafault TRUE.}

\item{write.format}{The format to write the concatenate alignment (only if write.dna == TRUE). Three options: 1. nexus (default), 2. phylip, 3. fasta or 4. tnt.}

\item{filename}{The name of the output file (only if write.dna==TRUE).}

\item{genes}{A list object with multiple alignments (DNAbin class)}
}
\value{
A concatenate alignment of DNAbin class
}
\description{
This function concatenate multiple alignments
}
\examples{

library(alignTools)

data(Procyonidae)
head(Procyonidae, 5L)

genes <- multiGenBank(Procyonidae,TRUE,FALSE)

alignment <- multiMuscle(x=genes, FALSE)

concat <- concatGenes(genes=alignment, missing="?", write.dna=TRUE, write.format="phylip", filename="Concat_eg_output")

str(concat)

}
\seealso{
\code{\link{multiMuscle}} \code{\link{completeGenes}}
}
\author{
Omar Daniel Leon-Alvarado \href{mailto:leon.alvarado12@gmail.com}{leon.alvarado12@gmail.com}
}
\keyword{alignments}
\keyword{concatenate}
\keyword{genes}
